refactor: move mint-card component out of dir
